created video player
