+++ lib/views/home/home_viewmodel.dart
@@ -1,7 +1,3 @@
-import 'package:ecngnews/models/news_category.dart';
-import 'package:ecngnews/models/news_model.dart';
-import 'package:ecngnews/services/news_service.dart';
-import 'package:ecngnews/utils/locator.dart';
 import 'package:stacked/stacked.dart';
 
 class HomeViewModel extends BaseViewModel {}
