+++ lib/views/watch/watch_view.dart
@@ -10,75 +10,8 @@ class WatchView extends StatefulWidget {
 }
 
 class _WatchViewState extends State<WatchView> {
-  YoutubePlayerController _youtubePlayerController;
-  @override
-  void initState() {
-    _youtubePlayerController = YoutubePlayerController(
-      initialVideoId: '1Rkn6rnsgc4',
-      flags: YoutubePlayerFlags(
-        mute: false,
-        autoPlay: false,
-        disableDragSeek: false,
-        loop: false,
-        isLive: false,
-        forceHideAnnotation: true,
-        forceHD: false,
-        enableCaption: true,
-        controlsVisibleAtStart: true,
-      ),
-    );
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _youtubePlayerController.dispose();
-
-    super.dispose();
-  }
-
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      children: <Widget>[
-        Container(
-          padding: EdgeInsets.only(
-              top: SizeConfig.heightMultiplier * 15,
-              left: SizeConfig.sizeMultiplier * 1.5,
-              right: SizeConfig.sizeMultiplier * 1.5,
-              bottom: SizeConfig.sizeMultiplier * 1.5),
-          child: ListView(
-            children: <Widget>[
-              Container(
-                height: SizeConfig.heightMultiplier * 30,
-//                    padding: EdgeInsets.all(10),
-                decoration: EcngTheme.cardDecoration,
-                child: Stack(
-                  children: <Widget>[
-                    Align(
-                      child: YoutubePlayer(
-                        width: SizeConfig.widthMultiplier * 90,
-                        bottomActions: <Widget>[
-                          Icon(Icons.play_arrow),
-                          Icon(Icons.pause),
-                        ],
-                        controller: _youtubePlayerController,
-                      ),
-                      alignment: Alignment.topCenter,
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-        Align(
-          alignment: AlignmentDirectional.topStart,
-          child: NewsCategoriesPan(
-            isDecorated: false,
-          ),
-        ),
-      ],
-    );
+    return Container();
   }
 }
